Refactor dfx orbit integration tests
