+++ tests/integration/src/dfx_orbit/canister_call.rs
@@ -1,5 +1,5 @@
 use crate::{
-    dfx_orbit::{
+    dfx_orbit::setup::{
         dfx_orbit_test, setup_counter_canister, setup_dfx_orbit, setup_dfx_user, DfxOrbitTestConfig,
     },
     setup::{setup_new_env, WALLET_ADMIN_USER},
